+++ test/shared/confirmation_message_test.dart
@@ -1,15 +1,16 @@
-import 'package:didpay/shared/async/async_data_widget.dart';
+import 'package:didpay/shared/confirmation_message.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../helpers/widget_helpers.dart';
+import '../helpers/widget_helpers.dart';
 
 void main() {
-  group('AsyncDataWidget', () {
+  group('ConfirmationMessage', () {
     testWidgets('should show request was submitted', (tester) async {
       await tester.pumpWidget(
         WidgetHelpers.testableWidget(
-          child: const AsyncDataWidget(text: 'Your request was submitted!'),
+          child:
+              const ConfirmationMessage(message: 'Your request was submitted!'),
         ),
       );
 
@@ -20,8 +21,8 @@ void main() {
     testWidgets('should show done button', (tester) async {
       await tester.pumpWidget(
         WidgetHelpers.testableWidget(
-          child: const AsyncDataWidget(
-            text: '',
+          child: const ConfirmationMessage(
+            message: '',
           ),
         ),
       );
